Refactor navigator tail and planner architecture
